+++ e2e/src/object/connections_store.rs
@@ -3,7 +3,10 @@
 
 use crate::{
     browser::Statement,
-    object::{Error, Object, connection::Connection},
+    object::{
+        Error, Object,
+        connection::{Connection, MemberConnectionState},
+    },
 };
 
 /// Storage for [`Connection`]s thrown by `Room.on_new_connection()` callback.
@@ -65,6 +68,85 @@ impl Object<ConnectionStore> {
             }
             ",
             [remote_id.into()],
+        ))
+        .await
+    }
+
+    /// Returns promise that resolves when [`MemberConnectionState`] is
+    /// in specific state.
+    ///
+    /// # Errors
+    ///
+    /// If failed to execute JS statement.
+    pub async fn wait_for_state(
+        &self,
+        remote_id: String,
+        state: Option<MemberConnectionState>,
+    ) -> Result<Object<()>, Error> {
+        self.execute_and_fetch(Statement::new(
+            // language=JavaScript
+            "
+            async (store) => {
+                const kinds = {
+                    P2P: window.rust.MemberConnectionStateKind.P2P,
+                };
+                const values = {
+                    P2P: {
+                        New: window.rust.PeerConnectionState.New,
+                        Connecting: window.rust.PeerConnectionState.Connecting,
+                        Connected: window.rust.PeerConnectionState.Connected,
+                        Disconnected:
+                            window.rust.PeerConnectionState.Disconnected,
+                        Failed: window.rust.PeerConnectionState.Failed,
+                        Closed: window.rust.PeerConnectionState.Closed,
+                    },
+                };
+
+                const [remoteId, expected] = args;
+                const connection = store.connections.get(remoteId);
+
+                if (!connection) {
+                    throw new NotFoundError();
+                }
+
+                const handle = connection.conn;
+                const state = handle.get_state();
+
+                if (state !== undefined) {
+                    if (expected === null) {
+                        throw new Error('State must not be set');
+                    }
+
+                    if (state.kind() !== kinds[expected.kind]) {
+                        throw new Error('Wrong MemberConnectionStateKind');
+                    }
+
+                    const value = values[expected.kind][expected.value];
+                    if (state.value() === value) {
+                        return;
+                    }
+                } else if (expected === null) {
+                    return;
+                }
+
+                return new Promise((resolve) => {
+                    connection.stateListener.subs.push((state) => {
+                        if (state.kind() !== kinds[expected.kind]) {
+                            return reject();
+                        }
+
+                        const value = values[expected.kind][expected.value];
+                        if (state.value() === value) {
+                            return resolve();
+                        }
+                    });
+                });
+            }
+            ",
+            [
+                remote_id.into(),
+                serde_json::to_value(state).map_err(|_e| Error::TypeCast)?,
+            ],
         ))
         .await
     }